RequestTargetを切り出し
